Abstracted repositories, moved controller/ out of service/ and created helper/
