+++ internal/helper/configLoader.go
@@ -1,4 +1,4 @@
-package clients
+package helper
 
 import (
 	"fmt"
