+++ internal/repository/notification.go
@@ -5,13 +5,17 @@ import (
 	"notification-service.com/packages/internal/dto"
 )
 
-type notification struct { }
-
-func NewNotificationRepository() *notification {
-	return &notification{}
+type NotificationRepository interface {
+	Insert(snr *dto.SendNotificationRequest, message *string) bool
 }
 
-func (n *notification) Insert(snr *dto.SendNotificationRequest, message *string) bool {
+type basicNotificationRepository struct { }
+
+func NewNotificationRepository() NotificationRepository {
+	return &basicNotificationRepository{}
+}
+
+func (bnr *basicNotificationRepository) Insert(snr *dto.SendNotificationRequest, message *string) bool {
 	stmt, err1 := clients.SQLClient.Prepare("insert into Notifications(Title, ContactType, ContactInfo, Message, UserId, AppId) values(?, ?, ?, ?, ?, ?)")
 	if err1 != nil {
 		return false
